requirements.txt
aioredis==1.3.1
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-restframework==0.0.1
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
isort==5.7.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.0
ocpp==0.8.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pylint==2.6.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
